move buttons.js into modules/
I think it makes more sense there since it's really part of the game
engine
